Still in progress, only pageparser still to do
Moved some code to new files
