--- smmry.js
@@ -1,6 +1,0 @@
-/**
- * Contains config info for querying the SMMRY API
- */
-
-export const smmryurl = "https://api.smmry.com/";
-export const apikey = "D7F33A666C";
